feat(multiline-cf): move markdown value login in getCustomFieldDisplayValue
