+++ app/routes/api+/utils.parse-markdown.ts
@@ -1,7 +1,7 @@
 import { json, type ActionFunctionArgs } from "@remix-run/node";
 import { makeShelfError } from "~/utils/error";
 import { assertIsPost, data, error } from "~/utils/http.server";
-import { parseMarkdownToReact } from "~/utils/md.server";
+import { parseMarkdownToReact } from "~/utils/md";
 
 export async function action({ context, request }: ActionFunctionArgs) {
   const authSession = context.getSession();
